[refac] 기존 Mui 테마 코드를 themeConfig.ts로 이동
