+++ front/src/components/Layout/Layout.tsx
@@ -1,24 +1,9 @@
-import { ThemeProvider, createTheme } from "@mui/material/styles";
+import { ThemeProvider } from "@mui/material/styles";
 import { Outlet } from "react-router-dom";
+import theme from "../../theme/themeConfig";
 import Header from "./Header";
 import LeftSideBar from "./LeftSideBar";
 const Layout: React.FC = () => {
-  const theme = createTheme({
-    palette: {
-      primary: {
-        light: "#FDE2F3",
-        main: "#917FB3",
-        dark: "#FDE2F3",
-        contrastText: "#000",
-      },
-      secondary: {
-        light: "#2A2F4F",
-        main: "#2A2F4F",
-        dark: "#FDE2F3",
-        contrastText: "#fff",
-      },
-    },
-  });
   return (
     <div>
       <Header />
