+++ front/src/pages/Login.tsx
@@ -1,22 +1,7 @@
-import { ThemeProvider, createTheme } from "@mui/material/styles";
+import { ThemeProvider } from "@mui/material/styles";
 import LoginBox from "../components/Auth/LoginBox";
+import theme from "../theme/themeConfig";
 const Login = () => {
-  const theme = createTheme({
-    palette: {
-      primary: {
-        light: "#FDE2F3",
-        main: "#917FB3",
-        dark: "#E5BEEC",
-        contrastText: "#fff",
-      },
-      secondary: {
-        light: "#2A2F4F",
-        main: "#2A2F4F",
-        dark: "#E5BEEC",
-        contrastText: "#fff",
-      },
-    },
-  });
   return (
     <div>
       <ThemeProvider theme={theme}>
