+++ front/src/theme/themeConfig.ts
@@ -1,0 +1,20 @@
+import { createTheme } from "@mui/material/styles";
+
+const theme = createTheme({
+  palette: {
+    primary: {
+      light: "#FDE2F3",
+      main: "#917FB3",
+      dark: "#E5BEEC",
+      contrastText: "#fff",
+    },
+    secondary: {
+      light: "#2A2F4F",
+      main: "#2A2F4F",
+      dark: "#E5BEEC",
+      contrastText: "#fff",
+    },
+  },
+});
+
+export default theme;
